+++ Lumina/Data/Structs/SqPackIndexHeader.cs
@@ -36,6 +36,6 @@ namespace Lumina.Data.Structs
 
         public byte DataFileId => (byte) ( ( data & 0b1110 ) >> 1 );
 
-        public uint Offset => (uint) ( data & ~0xF ) * 0x08;
+        public long Offset => (uint) ( data & ~0xF ) * 0x08;
     }
 }
